feat: choose a debate's join gate at creation
The identity registry was a deployment-wide immutable, so every debate on a deployment
answered to the same gate. It now sits on the debate, chosen by its creator, and join
consults that one.

Three modes fall out of the interface unchanged. The zero address leaves a debate open
to everyone - cheaper than a registry that always says yes, and it cannot later answer
differently. Any other address is asked isRegistered, which is how a personhood proof, a
curated allowlist and a group's membership all become the same gate. One registry serves
any number of debates, so a creator maintaining a group reuses it by address.

Deliberate now takes no constructor arguments, which is why the deploy script and the
verify recipes stop reconstructing them, and why deploy-with-mock is gone: the open mode
covers what it was for.

feePercentage narrows to uint8 in the same change. Its domain is 0..99 and the type now
says so; it also keeps the debate struct inside one slot with the registry added. The
narrowing removes conversions rather than adding them - uint8 widens implicitly into the
market math, so the fee reaches multiplyByFraction with no cast at all.

+++ src/libs/Parameters.sol
@@ -15,7 +15,7 @@ library Parameters {
     /// @notice The highest market fee (in percent) a debate creator may set. Capping below 100 keeps
     /// every nonzero stake's net amount at least 1, so a stake can never degenerate into a pure fee
     /// transfer that moves no market.
-    uint32 internal constant _MAX_FEE_PERCENTAGE = 99;
+    uint8 internal constant _MAX_FEE_PERCENTAGE = 99;
 
     /// @notice The initial vote token balance granted to a user upon joining a debate.
     uint32 public constant INITIAL_TOKENS = 100;
